feat(913): declared program compact card refacto
