+++ src/features/student/personalCareer/components/cards/DeclaredProgramCompactCard/DeclaredProgramCompactCard.stub.ts
@@ -2,6 +2,6 @@ import type { Component } from 'vue'
 
 export const DeclaredProgramCompactCardStub: Component = defineComponent({
   name: 'DeclaredProgramCompactCard',
-  props: ['title', 'valorized', 'iconName'],
+  props: ['title', 'valorized'],
   template: '<div data-testid="declared-program-compact-card-stub"></div>'
 })
